Conflicts:
	README.md

+++ ModelCatalogueCorePlugin/test/integration/org/modelcatalogue/core/AbstractControllerIntegrationSpec.groovy
@@ -1,12 +1,15 @@
 package org.modelcatalogue.core
 
 import grails.rest.RestfulController
+import grails.util.GrailsNameUtils
 import groovy.util.slurpersupport.GPathResult
+import org.apache.poi.xssf.usermodel.XSSFWorkbook
 import org.codehaus.groovy.grails.web.json.JSONElement
 import org.codehaus.groovy.grails.web.json.JSONObject
 import org.modelcatalogue.core.util.DefaultResultRecorder
 import org.modelcatalogue.core.util.Elements
 import org.modelcatalogue.core.util.ResultRecorder
+import org.modelcatalogue.core.util.marshalling.xlsx.XLSXListRenderer
 import spock.lang.Shared
 import spock.lang.Unroll
 
@@ -25,7 +28,6 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     def setupSpec(){
         loadMarshallers()
         loadFixtures()
-        RelationshipType.initDefaultRelationshipTypes()
         recorder = DefaultResultRecorder.create(
                 "../ModelCatalogueCorePlugin/target/xml-samples/modelcatalogue/core",
                 "../ModelCatalogueCorePlugin/test/js/modelcatalogue/core",
@@ -34,16 +36,12 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
         totalCount = 12
     }
 
-    def cleanupSpec(){
-
-    }
-
 
     @Unroll
     def "list json items test: #no where max: #max offset: #offset"() {
 
         expect:
-        resource.count() == total
+        resource.count() == totalCount
 
         when:
         controller.response.format = "json"
@@ -104,18 +102,35 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
         [no, size, max, offset, total, next, previous] << getPaginationParameters("/${resourceName}/")
     }
 
+    def "Export items to excel test"() {
+        controller.response.format = "xlsx"
+        controller.index()
+
+        XSSFWorkbook workbook = new XSSFWorkbook(new ByteArrayInputStream(controller.response.contentAsByteArray))
+
+        expect:
+        controller.response.contentType == XLSXListRenderer.EXCEL.name
+        workbook
+        workbook.getSheetAt(workbook.getActiveSheetIndex()).getLastRowNum() == totalRowsExported
+        // TODO: read the config and test the right number of columns as well
+    }
+
+    protected getTotalRowsExported() {
+        totalCount
+    }
+
     def "Show single existing item as JSON"() {
 
         when:
         controller.response.format = "json"
-        controller. params.id = "${loadItem.id}"
+        controller.params.id = "${loadItem.id}"
         controller.show()
         JSONObject json = controller.response.json
         recordResult 'showOne', json
 
         then:
         json
-        json.link == "/${resourceName}/${loadItem.id}"
+        json.link == "/${GrailsNameUtils.getPropertyName(loadItem.class)}/${loadItem.id}"
         customJsonPropertyCheck loadItem, json
         resource.count() == totalCount
 
@@ -134,7 +149,7 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
         then:
         xml
         xml.@link
-        xml.@link.text() == "/${resourceName}/${loadItem.id}"
+        xml.@link.text() == "/${GrailsNameUtils.getPropertyName(loadItem.class)}/${loadItem.id}"
         xmlCustomPropertyCheck xml, loadItem
         resource.count() == totalCount
 
@@ -143,6 +158,7 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
 
     @Unroll
     def "Do not #action new instance from JSON with bad json name"() {
+        if (controller.readOnly) return
 
         expect:
         !resource.findByName(badInstance.name)
@@ -169,6 +185,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "Create new instance from JSON"() {
+        if (controller.readOnly) return
+
         expect:
         !resource.findByName(newInstance.name)
 
@@ -190,8 +208,9 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
         resource.count() == totalCount
     }
 
-
     def "Create new instance from XML"() {
+        if (controller.readOnly) return
+
         expect:
         !resource.findByName(newInstance.name)
 
@@ -217,6 +236,7 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "edit instance description from JSON"() {
+        if (controller.readOnly) return
 
         def json = propertiesToEdit
         def properties = new HashMap()
@@ -246,6 +266,7 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "edit instance from XML"() {
+        if (controller.readOnly) return
 
         def properties = new HashMap()
         properties.putAll(loadItem.properties)
@@ -274,6 +295,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "Do not create new instance with bad XML"() {
+        if (controller.readOnly) return
+
         expect:
         !resource.findByName("")
 
@@ -298,6 +321,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "edit instance with bad JSON name"() {
+        if (controller.readOnly) return
+
         def instance = resource.findByName(loadItem.name)
 
         expect:
@@ -322,6 +347,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "edit instance with bad XML"() {
+        if (controller.readOnly) return
+
         def instance = resource.findByName(loadItem.name)
 
         expect:
@@ -367,6 +394,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "Return 404 for non-existing item as JSON on delete"() {
+        if (controller.readOnly) return
+
         controller.response.format = "json"
         controller.params.id = "1000000"
         controller.delete()
@@ -378,6 +407,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "Return 404 for non-existing item as XML on delete"() {
+        if (controller.readOnly) return
+
         controller.response.format = "xml"
         controller.params.id = "1000000"
         controller.params.id = "1000000"
@@ -391,6 +422,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
 
 
     def "Return 204 for existing item as JSON on delete"() {
+        if (controller.readOnly) return
+
         def elementToDelete = resource.newInstance(newInstance).save()
         controller.response.format = "json"
         controller.params.id = elementToDelete.id
@@ -404,6 +437,8 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def "Return 204 for existing item as XML on delete"() {
+        if (controller.readOnly) return
+
         def elementToDelete = resource.newInstance(newInstance).save()
         controller.response.format = "xml"
         controller.params.id = elementToDelete.id
@@ -533,7 +568,7 @@ abstract class AbstractControllerIntegrationSpec<T> extends AbstractIntegrationS
     }
 
     def checkPropertyMapMapString(Map property, Map value, String propertyName) {
-        if (property != value && (property!="" && value!=null)) {
+        if (property!= value && (property!="" && value!=null)) {
             throw new AssertionError("error: property to check: ${propertyName}  where property: ${property} !=  item: ${value}")
         }
     }
